fix: getNicknameByUserId, getEmailByuserId 메서드를 UserQueryService로 옮김
